+++ Sources/InterposeKit/InterposeError.swift
@@ -30,6 +30,9 @@ public enum InterposeError: LocalizedError {
     /// Core Foundation-backed objects do not support isa-swizzling from Swift.
     case coreFoundationObjectDetected(AnyObject)
 
+    /// A weakly referenced object deallocated before an operation that requires it.
+    case objectDeallocated
+
     /// Object is lying about it's actual class metadata.
     /// This usually happens when other swizzling libraries (like Aspects) also interfere with a class.
     /// While this might just work, it's not worth risking a crash, so similar to KVO this case is rejected.
@@ -73,6 +76,8 @@ extension InterposeError: Equatable {
             return "Unable to hook object that uses Key Value Observing: \(obj)"
         case .coreFoundationObjectDetected(let obj):
             return "Unable to hook Core Foundation-backed object: \(obj)"
+        case .objectDeallocated:
+            return "Unable to hook an object that has been deallocated"
         case .objectPosingAsDifferentClass(let obj, let actualClass):
             return "Unable to hook \(type(of: obj)) posing as \(NSStringFromClass(actualClass))/"
         case .invalidState(let expectedState):